Suggested Changes
These files include the vanilla Galactic Community Resolution (Extra-dimensional Experimentation) planetary decision and job that were overlooked in the building files.  The Delta tier building isn't included here.

Also included here are suggested changes to the resources, benefits and upkeep costs.  In my opinion, the economy is rendered to be completely irrelevant in the late game in terms of it being an aspect the player is concerned about.  The economic bonuses are simply to large for the player to care about it when these tiers are reached.  I suggest taking a look at the delta, phanon and omega tiers as well in this regard, because I think it's unfortunate that the economy becomes irrelevant, with no need to manage it any longer.

Apart from the galactic resolution stuff, my changes are annotated.  You can search #Mod Pack: to find what I've done, or simply compare the files with w/e programme you have for that.